refactor(web): extract MCP authorization orchestration
